backend_consorcio_automovel/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-ninja==1.3.0
pillow==10.4.0
pydantic==2.9.1
pydantic_core==2.23.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
